[feat] Smoke test + calibration + auto_lab + skills + lightweight reward
- Fix Qwen3VL loading (class, layer path, bf16→numpy)
- Smoke test: 9/9 pass, vision activation Δ confirmed non-trivial (mean=6.1)
- Calibration: 20 vision heads identified (top: L5H0 d=9.80)
- Auto lab state machine for experiment automation
- Lightweight InSituVisionReward with per-step hook capture
- Skills directory with 4 documented findings

+++ scripts/calibrate.py
@@ -36,7 +36,8 @@ def main():
     args = parser.parse_args()
 
     # Load model
-    model_info = load_model(args.model)
+    import torch
+    model_info = load_model(args.model, dtype=torch.float16)
 
     # Load calibration data
     samples = build_calibration_set(limit_per_source=args.limit_per_source)
